chore: move icons from sdk to dogfood
