+++ packages/react-native-sdk/src/icons/index.tsx
@@ -6,8 +6,6 @@ export * from './PhoneDown';
 export * from './Settings';
 export * from './Video';
 export * from './VideoSlash';
-export * from './Chat';
-export * from './Participants';
 export * from './ThreeDots';
 export * from './PinVertical';
 export * from './Spotlight';
@@ -21,7 +19,4 @@ export * from './StartStreamIcon';
 export * from './StopScreenShare';
 export * from './EndStreamIcon';
 export * from './LeaveStreamIcon';
-export * from './More';
 export * from './IconWrapper';
-export * from './Audio';
-export * from './RecordCall';
