+++ sample-apps/react-native/dogfood/src/assets/Audio.tsx
@@ -7,7 +7,7 @@ type Props = {
   size: number;
 };
 
-export const Audio = ({ color, size }: Props) => (
+const Audio = ({ color, size }: Props) => (
   <Svg viewBox={`0 0 ${size} ${size}`} width={size} height={size}>
     <Path
       d="M3 10.0018V14.0018C3 14.5518 3.45 15.0018 4 15.0018H7L10.29 18.2918C10.92 18.9218 12 18.4718 12 17.5818V6.41179C12 5.52179 10.92 5.07179 10.29 5.70179L7 9.00179H4C3.45 9.00179 3 9.45179 3 10.0018ZM16.5 12.0018C16.5 10.2318 15.48 8.71179 14 7.97179V16.0218C15.48 15.2918 16.5 13.7718 16.5 12.0018ZM14 4.45179V4.65179C14 5.03179 14.25 5.36179 14.6 5.50179C17.18 6.53179 19 9.06179 19 12.0018C19 14.9418 17.18 17.4718 14.6 18.5018C14.24 18.6418 14 18.9718 14 19.3518V19.5518C14 20.1818 14.63 20.6218 15.21 20.4018C18.6 19.1118 21 15.8418 21 12.0018C21 8.16179 18.6 4.89179 15.21 3.60179C14.63 3.37179 14 3.82179 14 4.45179Z"
@@ -15,3 +15,5 @@ export const Audio = ({ color, size }: Props) => (
     />
   </Svg>
 );
+
+export default Audio;
