+++ sample-apps/react-native/dogfood/src/assets/MoreActions.tsx
@@ -7,7 +7,7 @@ type Props = {
   size: number;
 };
 
-export const More = ({ color, size }: Props) => (
+const MoreActions = ({ color, size }: Props) => (
   <Svg viewBox={`0 0 ${size} ${size}`} width={size} height={size}>
     <Path
       d="M14.367 9.33329C15.6504 9.33329 16.7004 8.28329 16.7004 6.99996C16.7004 5.71663 15.6504 4.66663 14.367 4.66663C13.0837 4.66663 12.0337 5.71663 12.0337 6.99996C12.0337 8.28329 13.0837 9.33329 14.367 9.33329ZM14.367 11.6666C13.0837 11.6666 12.0337 12.7166 12.0337 14C12.0337 15.2833 13.0837 16.3333 14.367 16.3333C15.6504 16.3333 16.7004 15.2833 16.7004 14C16.7004 12.7166 15.6504 11.6666 14.367 11.6666ZM14.367 18.6666C13.0837 18.6666 12.0337 19.7166 12.0337 21C12.0337 22.2833 13.0837 23.3333 14.367 23.3333C15.6504 23.3333 16.7004 22.2833 16.7004 21C16.7004 19.7166 15.6504 18.6666 14.367 18.6666Z"
@@ -15,3 +15,5 @@ export const More = ({ color, size }: Props) => (
     />
   </Svg>
 );
+
+export default MoreActions;
