+++ sample-apps/react-native/dogfood/src/components/CallControlls/ChatButton.tsx
@@ -3,8 +3,9 @@ import {
   CallControlsButton,
   useTheme,
 } from '@stream-io/video-react-native-sdk';
-import { Chat, IconWrapper } from '@stream-io/video-react-native-sdk/src/icons';
+import { IconWrapper } from '@stream-io/video-react-native-sdk/src/icons';
 import { BadgeCountIndicator } from './BadgeCountIndicator';
+import Chat from '../../assets/Chat';
 
 /**
  * The props for the Chat Button in the Call Controls.
